// ai-character-ui/src/types/character.types.ts
export interface Character {
    _id: string;

    name: string;

    image?: string;

    tagline?: string;

    personality?: string;

    background?: string;

    emotionalBoundaries?: string;

    speechStyle?: string;

    relationshipStyle?: string;

    createdAt?: string;

    updatedAt?: string;
}